backend/requirements.txt
boto3>=1.34.0
python-geohash>=0.8.5
Pillow>=10.0.0
